requirements.txt
affine==2.4.0
attrs==24.2.0
blinker==1.5.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
greenlet==3.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
mariadb==1.1.12
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
PyMySQL==1.1.1
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
rasterio==1.4.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.40
threadpoolctl==3.5.0
tifffile==2025.3.30
typing_extensions==4.13.2
tzdata==2024.2
virtualenv==20.30.0
Werkzeug==3.1.3
